+++ packages/admin-api/src/clients/SheetClient.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@halo-dev/rest-api-client'
+import { HttpClient } from '@mtraders/rest-api-client'
 import { buildPath } from '../url'
 import {
   BasePostDetail,